Deprecated Swift 1.0 syntax.
Moved Swift 1.0 syntax & String extension to separate files.
