+++ Sources/Localize.swift
@@ -20,70 +20,6 @@ let LCLBaseBundle = "Base"
 /// Name for language change notification
 public let LCLLanguageChangeNotification = "LCLLanguageChangeNotification"
 
-// MARK: Localization Syntax
-
-/**
-Swift 1.x friendly localization syntax, replaces NSLocalizedString
-- Parameter string: Key to be localized.
-- Returns: The localized string.
-*/
-public func Localized(_ string: String) -> String {
-    return string.localized()
-}
-
-/**
- Swift 1.x friendly localization syntax with format arguments, replaces String(format:NSLocalizedString)
- - Parameter string: Key to be localized.
- - Returns: The formatted localized string with arguments.
- */
-public func Localized(_ string: String, arguments: CVarArg...) -> String {
-    return String(format: string.localized(), arguments: arguments)
-}
-
-/**
- Swift 1.x friendly plural localization syntax with a format argument
- 
- - parameter string:   String to be formatted
- - parameter argument: Argument to determine pluralisation
- 
- - returns: Pluralized localized string.
- */
-public func LocalizedPlural(_ string: String, argument: CVarArg) -> String {
-    return string.localizedPlural(argument)
-}
-
-
-public extension String {
-    /**
-     Swift 2 friendly localization syntax, replaces NSLocalizedString
-     - Returns: The localized string.
-     */
-    func localized() -> String {
-        return localized(using: nil, in: .main)
-    }
-
-    /**
-     Swift 2 friendly localization syntax with format arguments, replaces String(format:NSLocalizedString)
-     - Returns: The formatted localized string with arguments.
-     */
-    func localizedFormat(_ arguments: CVarArg...) -> String {
-        return String(format: localized(), arguments: arguments)
-    }
-    
-    /**
-     Swift 2 friendly plural localization syntax with a format argument
-     
-     - parameter argument: Argument to determine pluralisation
-     
-     - returns: Pluralized localized string.
-     */
-    func localizedPlural(_ argument: CVarArg) -> String {
-        return NSString.localizedStringWithFormat(localized() as NSString, argument) as String
-    }
-}
-
-
-
 // MARK: Language Setting Functions
 
 open class Localize: NSObject {
